Updat approximate infeasibility criteria.
This changes the approximate infeasibility criteria for FirstOrderLp to be:
  primal infeasibility:
    (1) dual_ray_objective > 0.0 (was >= eps_primal_infeasible),
    (2) max_dual_ray_infeasibility / dual_ray_objective <=
        eps_primal_infeasible (was <).
  dual infeasibility:
    (1) primal_ray_linear_objective < 0.0 (was <= eps_dual_infeasible),
    (2) max_primal_ray_infeasibility / (-primal_ray_linear_objective) <=
        eps_dual_infeasible (was <),
    (3) primal_ray_quadratic_norm / (-primal_ray_linear_objective) <=
        eps_dual_infeasible (was <).

The motivation for changing (1) is primarily that for a non-zero epsilon,
there are simple instances that converge to a ray_objective between 0 and
epsilon, so that no matter how long we run, FirstOrderLp would never be able to
terminate. It also eliminates an annoying non-monoticity for infeasibility with
respect to epsilon.

The motivation for changing (2) is just that it allows 0.0 to be a meaningful
value for epsilon.  Otherwise a precisely feasible ray would be rejected by
epsilon == 0.0.  This was actually hit by one of the test cases ("Infeasible
instance" for PDHG, "Primal infeasible instance" for mirror prox) which used to
fail to detect infeasibility because it was using epsilon = 0.0, and hence
diverged, but now detects infeasibility.

TESTED=unit tests