Add netlib benchmark set.
This adds a script for collecting and building the netlib benchmark
set, as well as instructions in README.md. Because the netlib
benchmark instances use fixed-format MPS in a way that is, for some
instances, incompatible with free-format MPS readers, this also
includes a hacky awk script for converting fixed-format MPS files into
a version readable by free-format readers.